rename + limit navigation by route lifecycle
