+++ app/src/main/java/com/mapbox/navigation/examples/standalone/StandaloneExamplesList.kt
@@ -21,7 +21,7 @@ import com.mapbox.navigation.examples.standalone.speedlimit.ShowSpeedLimitActivi
 import com.mapbox.navigation.examples.standalone.status.ShowCustomStatusActivity
 import com.mapbox.navigation.examples.standalone.status.ShowStatusActivity
 import com.mapbox.navigation.examples.standalone.tripprogress.ShowTripProgressActivity
-import com.mapbox.navigation.examples.standalone.customforegroundservice.CustomForegroundServiceActivity
+import com.mapbox.navigation.examples.standalone.background.BackgroundExampleActivity
 import com.mapbox.navigation.examples.standalone.voice.PlayVoiceInstructionsActivity
 
 fun Context.examplesList() = listOf(
@@ -83,7 +83,7 @@ fun Context.examplesList() = listOf(
         ContextCompat.getDrawable(this, R.drawable.mapbox_screenshot_tbt_experience),
         getString(R.string.title_turn_by_turn),
         getString(R.string.description_turn_by_turn),
-        CustomForegroundServiceActivity::class.java
+        BackgroundExampleActivity::class.java
     ),
     MapboxExample(
         ContextCompat.getDrawable(this, R.drawable.mapbox_route_preview),
@@ -143,6 +143,6 @@ fun Context.examplesList() = listOf(
         ContextCompat.getDrawable(this, R.drawable.mapbox_screenshot_tbt_experience),
         getString(R.string.description_custom_foreground_service),
         getString(R.string.description_turn_by_turn),
-        CustomForegroundServiceActivity::class.java
+        BackgroundExampleActivity::class.java
     ),
 )
