+++ app/src/main/java/com/mapbox/navigation/examples/standalone/background/BackgroundExampleActivity.kt
@@ -1,23 +1,14 @@
-package com.mapbox.navigation.examples.standalone.customforegroundservice
+package com.mapbox.navigation.examples.standalone.background
 
 import android.annotation.SuppressLint
-import android.app.Notification
-import android.app.NotificationManager
-import android.app.PendingIntent
 import android.app.Service
 import android.content.Intent
 import android.content.res.Configuration
 import android.content.res.Resources
 import android.location.Location
-import android.os.Build
 import android.os.Bundle
 import android.os.IBinder
-import android.util.Log
-import android.view.View
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.NotificationChannelCompat
-import androidx.core.app.NotificationCompat
-import androidx.core.app.NotificationManagerCompat
 import androidx.core.content.ContextCompat
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
@@ -50,19 +41,19 @@ import com.mapbox.navigation.core.replay.route.ReplayProgressObserver
 import com.mapbox.navigation.core.replay.route.ReplayRouteMapper
 import com.mapbox.navigation.core.trip.session.LocationMatcherResult
 import com.mapbox.navigation.core.trip.session.LocationObserver
+import com.mapbox.navigation.core.trip.session.NavigationSessionState
+import com.mapbox.navigation.core.trip.session.NavigationSessionStateObserver
 import com.mapbox.navigation.core.trip.session.RouteProgressObserver
 import com.mapbox.navigation.core.trip.session.VoiceInstructionsObserver
 import com.mapbox.navigation.examples.R
-import com.mapbox.navigation.examples.databinding.MapboxActivityTurnByTurnExperienceBinding
+import com.mapbox.navigation.examples.databinding.MapboxActivityCustomForegroundServiceBinding
 import com.mapbox.navigation.ui.base.util.MapboxNavigationConsumer
 import com.mapbox.navigation.ui.maps.NavigationStyles
 import com.mapbox.navigation.ui.maps.camera.NavigationCamera
 import com.mapbox.navigation.ui.maps.camera.data.MapboxNavigationViewportDataSource
 import com.mapbox.navigation.ui.maps.camera.lifecycle.NavigationBasicGesturesHandler
-import com.mapbox.navigation.ui.maps.camera.state.NavigationCameraState
 import com.mapbox.navigation.ui.maps.camera.transition.NavigationCameraTransitionOptions
 import com.mapbox.navigation.ui.maps.location.NavigationLocationProvider
-import com.mapbox.navigation.ui.maps.route.arrow.model.RouteArrowOptions
 import com.mapbox.navigation.ui.maps.route.line.api.MapboxRouteLineApi
 import com.mapbox.navigation.ui.maps.route.line.api.MapboxRouteLineView
 import com.mapbox.navigation.ui.maps.route.line.model.MapboxRouteLineOptions
@@ -101,7 +92,7 @@ import java.util.Locale
  * - You can use buttons to mute/unmute voice instructions, recenter the camera, or show the route overview.
  */
 @OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
-class CustomForegroundServiceActivity : AppCompatActivity() {
+class BackgroundExampleActivity : AppCompatActivity() {
 
     private companion object {
         private const val BUTTON_ANIMATION_DURATION = 1500L
@@ -125,7 +116,7 @@ class CustomForegroundServiceActivity : AppCompatActivity() {
     /**
      * Bindings to the example layout.
      */
-    private lateinit var binding: MapboxActivityTurnByTurnExperienceBinding
+    private lateinit var binding: MapboxActivityCustomForegroundServiceBinding
 
     /**
      * Used to execute camera transitions based on the data generated by the [viewportDataSource].
@@ -231,6 +222,13 @@ class CustomForegroundServiceActivity : AppCompatActivity() {
         }
     }
 
+    private val navigationSessionStateObserver = NavigationSessionStateObserver { state ->
+        if (state is NavigationSessionState.ActiveGuidance) {
+            val serviceIntent = Intent(this, BackgroundService::class.java)
+            startService(serviceIntent)
+        }
+    }
+
     /**
      * Gets notified with progress along the currently active route.
      */
@@ -288,6 +286,7 @@ class CustomForegroundServiceActivity : AppCompatActivity() {
                 mapboxNavigation.registerLocationObserver(locationObserver)
                 mapboxNavigation.registerRouteProgressObserver(routeProgressObserver)
                 mapboxNavigation.registerRouteProgressObserver(replayProgressObserver)
+                mapboxNavigation.registerNavigationSessionStateObserver(navigationSessionStateObserver)
                 // start the trip session to being receiving location updates in free drive
                 // and later when a route is set also receiving route progress updates
                 mapboxNavigation.startTripSession(withForegroundService = true)
@@ -298,6 +297,7 @@ class CustomForegroundServiceActivity : AppCompatActivity() {
                 mapboxNavigation.unregisterLocationObserver(locationObserver)
                 mapboxNavigation.unregisterRouteProgressObserver(routeProgressObserver)
                 mapboxNavigation.unregisterRouteProgressObserver(replayProgressObserver)
+                mapboxNavigation.unregisterNavigationSessionStateObserver(navigationSessionStateObserver)
             }
         },
         onInitialize = this::initNavigation
@@ -306,7 +306,7 @@ class CustomForegroundServiceActivity : AppCompatActivity() {
     @SuppressLint("MissingPermission")
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = MapboxActivityTurnByTurnExperienceBinding.inflate(layoutInflater)
+        binding = MapboxActivityCustomForegroundServiceBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
         // initialize Navigation Camera
@@ -321,16 +321,6 @@ class CustomForegroundServiceActivity : AppCompatActivity() {
         binding.mapView.camera.addCameraAnimationsLifecycleListener(
             NavigationBasicGesturesHandler(navigationCamera)
         )
-        navigationCamera.registerNavigationCameraStateChangeObserver { navigationCameraState ->
-            // shows/hide the recenter button depending on the camera state
-            when (navigationCameraState) {
-                NavigationCameraState.TRANSITION_TO_FOLLOWING,
-                NavigationCameraState.FOLLOWING -> binding.recenter.visibility = View.INVISIBLE
-                NavigationCameraState.TRANSITION_TO_OVERVIEW,
-                NavigationCameraState.OVERVIEW,
-                NavigationCameraState.IDLE -> binding.recenter.visibility = View.VISIBLE
-            }
-        }
         // set the padding values depending on screen orientation and visible view layout
         if (this.resources.configuration.orientation == Configuration.ORIENTATION_LANDSCAPE) {
             viewportDataSource.overviewPadding = landscapeOverviewPadding
@@ -353,9 +343,6 @@ class CustomForegroundServiceActivity : AppCompatActivity() {
         routeLineApi = MapboxRouteLineApi(mapboxRouteLineOptions)
         routeLineView = MapboxRouteLineView(mapboxRouteLineOptions)
 
-        // initialize maneuver arrow view to draw arrows on the map
-        val routeArrowOptions = RouteArrowOptions.Builder(this).build()
-
         // load map style
         binding.mapView.getMapboxMap().loadStyleUri(NavigationStyles.NAVIGATION_DAY_STYLE) {
             // add long click listener that search for a route to the clicked destination
@@ -368,26 +355,7 @@ class CustomForegroundServiceActivity : AppCompatActivity() {
         // initialize view interactions
         binding.stop.setOnClickListener {
             clearRouteAndStopNavigation()
-            stopService(Intent(this, CustomForegroundService::class.java))
         }
-        binding.recenter.setOnClickListener {
-            navigationCamera.requestNavigationCameraToFollowing()
-            binding.routeOverview.showTextAndExtend(BUTTON_ANIMATION_DURATION)
-        }
-        binding.routeOverview.setOnClickListener {
-            navigationCamera.requestNavigationCameraToOverview()
-            binding.recenter.showTextAndExtend(BUTTON_ANIMATION_DURATION)
-        }
-
-        // set initial sounds button state
-        binding.soundButton.unmute()
-
-        val serviceIntent = Intent(this, CustomForegroundService::class.java)
-//        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-//            startForegroundService(serviceIntent)
-//        } else {
-            startService(serviceIntent)
-//        }
     }
 
     override fun onDestroy() {
@@ -415,7 +383,7 @@ class CustomForegroundServiceActivity : AppCompatActivity() {
             setLocationProvider(navigationLocationProvider)
             this.locationPuck = LocationPuck2D(
                 bearingImage = ContextCompat.getDrawable(
-                    this@CustomForegroundServiceActivity,
+                    this@BackgroundExampleActivity,
                     R.drawable.mapbox_navigation_puck_icon
                 )
             )
@@ -504,7 +472,7 @@ class CustomForegroundServiceActivity : AppCompatActivity() {
     }
 }
 
-class CustomForegroundService : Service(), LifecycleOwner {
+class BackgroundService : Service(), LifecycleOwner {
 
     private val lifecycleRegistry = LifecycleRegistry(this)
 
@@ -513,10 +481,12 @@ class CustomForegroundService : Service(), LifecycleOwner {
             @SuppressLint("MissingPermission")
             override fun onAttached(mapboxNavigation: MapboxNavigation) {
                 mapboxNavigation.registerVoiceInstructionsObserver(voiceInstructionsObserver)
+                mapboxNavigation.registerNavigationSessionStateObserver(navigationSessionStateObserver)
             }
 
             override fun onDetached(mapboxNavigation: MapboxNavigation) {
                 mapboxNavigation.unregisterVoiceInstructionsObserver(voiceInstructionsObserver)
+                mapboxNavigation.unregisterNavigationSessionStateObserver(navigationSessionStateObserver)
             }
         }
     )
@@ -614,5 +584,11 @@ class CustomForegroundService : Service(), LifecycleOwner {
             // remove already consumed file to free-up space
             speechApi.clean(value)
         }
+
+    private val navigationSessionStateObserver = NavigationSessionStateObserver { state ->
+        if (state !is NavigationSessionState.ActiveGuidance) {
+            stopSelf()
+        }
+    }
 
 }
