+++ app/src/main/java/com/mapbox/navigation/examples/standalone/background/res/values/strings.xml
@@ -2,5 +2,5 @@
 
     <string name="title_custom_foreground_service">Example of running navigation session in the background.</string>
     <string name="description_custom_foreground_service">The example starts a foreground service which keeps the session running until a user explicitly stops it.</string>
-
+    <string name="stop_navigation_session_button"></string>
 </resources>
